Introduce `gosec` for Static Application Security Testing (SAST) (#541)
* add make target to run gosec linter

* address gosec linter findings

* add gosec / linter logs to component descriptor

and drop no longer needed image builder

+++ pkg/apis/config/config_testmachinery.go
@@ -66,7 +66,7 @@ type GitHub struct {
 // GitHubCache is the github cache configuration
 type GitHubCache struct {
 	CacheDir        string `json:"cacheDir,omitempty"`
-	CacheDiskSizeGB int    `json:"cacheDiskSizeGB,omitempty"`
+	CacheDiskSizeGB uint64 `json:"cacheDiskSizeGB,omitempty"`
 	MaxAgeSeconds   int    `json:"maxAgeSeconds,omitempty"`
 }
